requirements.txt
customtkinter
Pillow
requests
beautifulsoup4
selenium
webdriver-manager
pystray
